+++ src/__tests__/client.test.js
@@ -40,14 +40,14 @@ test("can sign out", async () => {
 test("can call getProfile after logged in", async () => {
   const number = Math.random();
 
-  testServer.mock({ 
-    status: 200, 
-    body: { 
-      apiKey: { 
+  testServer.mock({
+    status: 200,
+    body: {
+      apiKey: {
         token: "baz",
         user: { number }
-      } 
-    } 
+      }
+    }
   });
   await signIn({ email: "foo", password: "bar" });
 
@@ -69,14 +69,14 @@ test("substitutes values into the request path", async () => {
 });
 
 test("can start a user simulation", async () => {
-  testServer.mock({ 
-    status: 200, 
-    body: { 
-      apiKey: { 
+  testServer.mock({
+    status: 200,
+    body: {
+      apiKey: {
         token: "bar",
         user: { id: 42 }
-      } 
-    } 
+      }
+    }
   });
 
   setToken("foo");